Instrumenation key can be set from code, and moved TelemetryClient and impl
