+++ Core/src/main/java/com/microsoft/applicationinsights/logging/common/TelemetryClientProxy.java
@@ -1,6 +1,6 @@
 package com.microsoft.applicationinsights.logging.common;
 
-import com.microsoft.applicationinsights.channel.TelemetryClient;
+import com.microsoft.applicationinsights.TelemetryClient;
 
 public interface TelemetryClientProxy {
