+++ src/main/java/com/microsoft/applicationinsights/tests/manual/Program.java
@@ -1,8 +1,9 @@
 package com.microsoft.applicationinsights.tests.manual;
 
-import com.microsoft.applicationinsights.channel.TelemetryClient;
-import com.microsoft.applicationinsights.datacontracts.*;
-import com.microsoft.applicationinsights.util.DefaultTelemetryClient;
+import com.microsoft.applicationinsights.TelemetryClient;
+import com.microsoft.applicationinsights.DefaultTelemetryClient;
+import com.microsoft.applicationinsights.datacontracts.MetricTelemetry;
+import com.microsoft.applicationinsights.datacontracts.HttpRequestTelemetry;
 
 import java.io.IOException;
 import java.util.Date;
